renamed __init__ and project
